+++ gcdt/gcdt_lifecycle.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals, print_function
-import copy
 import sys
 import logging
 
@@ -14,7 +13,6 @@ from . import gcdt_signals
 from .gcdt_defaults import DEFAULT_CONFIG
 from .utils import dict_merge, get_context, check_gcdt_update, \
     are_credentials_still_valid
-from .config_reader import read_config
 from .gcdt_cmd_dispatcher import cmd, get_command
 from .gcdt_plugins import load_plugins
 from .gcdt_awsclient import AWSClient
@@ -57,10 +55,11 @@ def lifecycle(awsclient, tool, command, arguments):
     gcdt_signals.initialized.send(context)
     check_gcdt_update()
 
-    gcdt_signals.config_read_init.send(context)
-    config = read_config(awsclient, config_base_name=
-        DEFAULT_CONFIG[tool].get('config_base_name', tool))
-    gcdt_signals.config_read_finalized.send(context)
+    config = {}
+    gcdt_signals.config_read_init.send((context, config))
+    gcdt_signals.config_read_finalized.send((context, config))
+
+    # TODO lookup
 
     gcdt_signals.config_validation_init.send((context, config))
     # TODO config validation
@@ -75,13 +74,20 @@ def lifecycle(awsclient, tool, command, arguments):
     are_credentials_still_valid(awsclient)
 
     # merge DEFAULT_CONFIG with config
-    tool_config = copy.deepcopy(DEFAULT_CONFIG[tool])
-    dict_merge(tool_config, config)
+    #tool_config = copy.deepcopy(DEFAULT_CONFIG[tool])
+    #dict_merge(tool_config, config)
+
+    # TODO bundle-phase if the tool & command requires one
+    gcdt_signals.bundle_init.send((context, config))
+    gcdt_signals._bundle.send((context, config))
+    gcdt_signals.bundle_finalized.send((context, config))
 
     # run the command and provide context and config (= tooldata)
     gcdt_signals.command_init.send((context, config))
     try:
-        exit_code = cmd.dispatch(arguments, context=context, config=config)
+        exit_code = cmd.dispatch(arguments,
+                                 context=context,
+                                 config=config[tool])
     except Exception as e:
         print(str(e))
         context['error'] = str(e)