+++ gcdt/route53.py
@@ -3,7 +3,8 @@
 """Helper to create Route53 entries
 """
 
-# FIXME: it looks like this is not used at all!!
+# TODO: this is probably only used in cloudformation
+# TODO: add documentation how to use it!
 
 from __future__ import unicode_literals, print_function
 import sys
@@ -12,7 +13,7 @@ import troposphere
 from troposphere.ec2 import Instance
 from troposphere.route53 import RecordSetType
 
-from .config_reader import get_env
+from .utils import get_env
 from .servicediscovery import get_outputs_for_stack
 
 TTL_DEFAULT = 300