+++ gcdt/tenkai_main.py
@@ -30,20 +30,27 @@ def version_cmd():
 @cmd(spec=['deploy'])
 def deploy_cmd(**tooldata):
     context = tooldata.get('context')
-    conf = tooldata.get('config')
+    config = tooldata.get('config')
     awsclient = context.get('_awsclient')
 
-    # are_credentials_still_valid()
     prepare_artifacts_bucket(awsclient,
-                             conf.get('codedeploy.artifactsBucket'))
+                             config['codedeploy'].get('artifactsBucket'))
+    # TODO deprecate prebundle hook with reference to new signal-based-hooks
+    pre_bundle_scripts = config.get('preBundle', None)
+    if pre_bundle_scripts:
+        exit_code = utils.execute_scripts(pre_bundle_scripts)
+        if exit_code != 0:
+            print('Pre bundle script exited with error')
+            return 1
+
     deployment = deploy(
         awsclient=awsclient,
-        applicationName=conf.get('codedeploy.applicationName'),
-        deploymentGroupName=conf.get('codedeploy.deploymentGroupName'),
-        deploymentConfigName=conf.get('codedeploy.deploymentConfigName'),
-        bucket=conf.get('codedeploy.artifactsBucket'),
-        pre_bundle_scripts=conf.get('preBundle', None)
+        applicationName=config['codedeploy'].get('applicationName'),
+        deploymentGroupName=config['codedeploy'].get('deploymentGroupName'),
+        deploymentConfigName=config['codedeploy'].get('deploymentConfigName'),
+        bucket=config['codedeploy'].get('artifactsBucket')
     )
+
     exit_code = deployment_status(awsclient, deployment)
     if exit_code:
         return 1
@@ -55,6 +62,7 @@ def bundle_cmd(**tooldata):
 
 
 def main():
+    # TODO: register bundle with bundle signals
     sys.exit(
         gcdt_lifecycle.main(DOC, 'tenkai',
                             dispatch_only=['version', 'bundle'])