+++ gcdt/utils.py
@@ -10,7 +10,6 @@ from clint.textui import prompt, colored
 from pyhocon import ConfigFactory
 
 from gcdt import __version__
-#from gcdt.config_reader import _get_datadog_api_key
 from gcdt.package_utils import get_package_versions
 
 
@@ -92,7 +91,7 @@ def _get_user():
     return getpass.getuser()
 
 
-def _get_env():
+def get_env():
     """
     Read environment from ENV and mangle it to a (lower case) representation
     :return: Environment as lower case string (or None if not matched)
@@ -121,7 +120,7 @@ def get_context(awsclient, tool, command, arguments=None):
         'user': _get_user()
     }
 
-    env = _get_env()
+    env = get_env()
     if env:
         context['env'] = env
 
@@ -198,26 +197,3 @@ def are_credentials_still_valid(awsclient):
         print(colored.red('Your credentials have expired... Please renew and try again!'))
         return 1
     return 0
-
-
-# dead code, not used!!
-'''
-def check_aws_credentials(awsclient):
-    """
-    A decorator that will check for valid credentials
-    """
-
-    def wrapper(func):
-        @wraps(func)
-        def wrapped(*args, **kwargs):
-            exit_code = are_credentials_still_valid(awsclient)
-            if exit_code:
-                # TODO: remove exit()
-                sys.exit()
-            result = func(*args, **kwargs)
-            return result
-
-        return wrapped
-
-    return wrapper
-'''