+++ tests/test_kumo_util.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals, print_function
 import troposphere
-from nose.tools import assert_equal
 from gcdt.kumo_util import StackLookup